requirements.txt
httpx==0.25.0  # For HTTP requests
pytest==7.4.2
python-dotenv==1.0.0
flake8==6.0.0
pytest-cov==4.1.0
